requirements.txt
PyYAML
skyscanner
pandas
